+++ src/check_jsonschema/loaders.py
@@ -2,9 +2,9 @@ import json
 import os
 import typing as t
 
-import identify
 import jsonschema
 import ruamel.yaml
+from identify import identify
 
 from .cachedownloader import CacheDownloader
 
@@ -45,8 +45,17 @@ class SchemaLoader:
 
 
 class InstanceLoader:
-    def __init__(self, filenames):
+    def __init__(self, filenames, default_filetype=None):
         self._filenames = filenames
+        self._default_ft = default_filetype
+
+    @property
+    def _default_loadfunc(self):
+        if not self._default_ft:
+            return None
+        if self._default_ft.lower() == "json":
+            return json.load
+        return yaml.load
 
     def iter_files(self):
         for fn in self._filenames:
@@ -56,9 +65,13 @@ class InstanceLoader:
             elif "json" in tags:
                 loadfunc = json.load
             else:
-                # TODO: handle this by storing it in the errors map
+                loadfunc = self._default_loadfunc
+
+            # TODO: handle this by storing it in the errors map
+            if not loadfunc:
                 raise BadFileTypeError(
                     f"cannot check {fn} as it is neither yaml nor json"
                 )
+
             with open(fn) as fp:
                 yield (fn, loadfunc(fp))